requirements.txt
async_timeout == 3.0.1
cexprtk == 0.3.4
discord.py == 1.5.1
passlib == 1.7.2
pyfiglet == 0.8.post1
rich == 6.0.0
sympy == 1.6.2
urbandict == 0.6.1
wikipedia == 1.4.0
youtube_dl == 2020.11.1.1
